chore(suite): rename firmware components, remove typography overrides
