+++ packages/suite/src/components/firmware/FirmwareButton.tsx
